refactor: refactors session handling to be simpler, makes the session storage layer opaque, and abstracts token handling away

+++ backend/api/src/constants/sessions.rs
@@ -1,6 +1,6 @@
 //! Constants related to authentication and session handling.
 
 /// Timeout for authenticated sessions in seconds.
-pub const SESSION_TIMEOUT: i64 = 7 * 24 * 60 * 60;
+pub const SESSION_TIMEOUT: u32 = 7 * 24 * 60 * 60;
 /// Timeout for authentication sessions in seconds.
-pub const AUTH_SESSION_TIMEOUT: i64 = 300;
+pub const AUTH_SESSION_TIMEOUT: u32 = 300;
